+++ projects/swimlane/ngx-ui/CHANGELOG.md
@@ -2,10 +2,12 @@
 
 # HEAD (unreleased)
 
+## 34.0.0 (2020-12-17)
+
 - Fix(DateTime): Component no longer emits a `(change)` event when input value is invalid.
 - Fix(DateTime): Display value not updated correctly.
-- Breaking: `ngx-plus-menu` takes items containing `title` and `subtitle`. Does not setup it's own hotkeys.
-- Fix: no popup us shown if `menuTitle` is not passed as an input to `ngx-plus-menu`
+- Breaking(ngx-plus-menu): items now contain `title` and `subtitle`. Does not setup it's own hotkeys.
+- Fix(ngx-plus-menu): no popup us shown if `menuTitle` is not passed as an input to `ngx-plus-menu`
 
 ## 33.1.0 (2020-12-14)